requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
python-dotenv
numpy
requests
python-telegram-bot
httpx
apscheduler==3.10.4
slowapi==0.1.9
# --- AI CORE UPGRADE ---
scikit-learn
pandas
joblib